Data inputs w/revised name changes
I added a new column that captures the abbreviated names.  

For CPI, I added one column with the CPI preceding it and another without.

For unemployment, earnings, and stocks, I'm assuming that we can add an an abbreviation for each drop down title.  For example: Earnings (E), Unemployment (U), and stocks (S).  So, each name starts with the abbreviation and name.  For example, earnings for all people would be "E - All People."  Happy to change this if needed.